# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Constituent Touchpoint Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{id}/touches:
    get:
      operationId: listTouches
      tags: [touches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Touch" }
    post:
      operationId: createTouch
      tags: [touches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TouchInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Touch" }
  /campaigns/{id}/touches/{touchId}:
    patch:
      operationId: updateTouch
      tags: [touches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: touchId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TouchUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Touch" }
    delete:
      operationId: deleteTouch
      tags: [touches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: touchId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ───────── Thresholds
  /campaigns/{id}/thresholds:
    get:
      operationId: listThresholds
      tags: [thresholds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ThresholdCheck" }
    post:
      operationId: createThreshold
      tags: [thresholds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ThresholdInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ThresholdCheck" }
  /campaigns/{id}/thresholds/{thresholdId}:
    delete:
      operationId: deleteThreshold
      tags: [thresholds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: thresholdId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /campaigns/{id}/threshold-preview:
    post:
      operationId: previewThresholds
      tags: [thresholds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ThresholdPreview" }
  /campaigns/{id}/threshold-overrides:
    post:
      operationId: setThresholdOverrides
      tags: [thresholds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OverrideInput" }
      responses:
        "204": { description: Saved }

  # ───────── Suppressions
  /campaigns/{id}/suppressions:
    get:
      operationId: listSuppressions
      tags: [suppressions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Suppression" }
    post:
      operationId: createSuppression
      tags: [suppressions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SuppressionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Suppression" }
  /campaigns/{id}/suppressions/{suppressionId}:
    delete:
      operationId: deleteSuppression
      tags: [suppressions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: suppressionId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ───────── Seeds
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppSettings" }
    patch:
      operationId: updateSettings
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppSettings" }
  /retention/delete:
    post:
      operationId: runRetentionDelete
      tags: [retention]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RetentionInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RetentionResult" }
components:
  schemas:
...
    Touch:
      type: object
      required: [id, campaignId, touchName, channelId, channelLabel, campaignTypeId, campaignTypeLabel, sendDate]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        touchName: { type: string }
        channelId: { type: integer }
        channelLabel: { type: string }
        campaignTypeId: { type: integer }
        campaignTypeLabel: { type: string }
        sendDate: { type: string, format: date }
        notes: { type: ["string", "null"] }
    TouchInput:
      type: object
      required: [touchName, channelId, campaignTypeId, sendDate]
      properties:
        touchName: { type: string }
        channelId: { type: integer }
        campaignTypeId: { type: integer }
        sendDate: { type: string, format: date }
        notes: { type: string }
    TouchUpdate:
      type: object
      properties:
        touchName: { type: string }
        channelId: { type: integer }
        campaignTypeId: { type: integer }
        sendDate: { type: string, format: date }
        notes: { type: string }
    ThresholdCheck:
      type: object
      required: [id, campaignId, name, maxTouchpoints, windowDays, scope, actionMode]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        name: { type: string }
        maxTouchpoints: { type: integer }
        windowDays: { type: integer }
        scope: { type: string, enum: [all, channel, campaign_type, channel_and_type] }
        channelId: { type: ["integer", "null"] }
        campaignTypeId: { type: ["integer", "null"] }
        actionMode: { type: string, enum: [track, flag, remove, manual] }
    ThresholdInput:
      type: object
      required: [name, maxTouchpoints, windowDays, scope, actionMode]
      properties:
        name: { type: string }
        maxTouchpoints: { type: integer, minimum: 1 }
        windowDays: { type: integer, minimum: 1 }
        scope: { type: string, enum: [all, channel, campaign_type, channel_and_type] }
        channelId: { type: integer }
        campaignTypeId: { type: integer }
        actionMode: { type: string, enum: [track, flag, remove, manual] }
    ThresholdConflict:
      type: object
      required: [donorId, thresholdId, thresholdName, projectedCount, maxAllowed, windowDays, explanation, overridden]
      properties:
        donorId: { type: string }
        thresholdId: { type: integer }
        thresholdName: { type: string }
        projectedCount: { type: integer }
        maxAllowed: { type: integer }
        windowDays: { type: integer }
        explanation: { type: string }
        overridden: { type: boolean }
    ThresholdPreview:
      type: object
      required: [conflicts, totalFlaggedDonors, totalProjectedTouchpoints, byThreshold]
      properties:
        conflicts:
          type: array
          items: { $ref: "#/components/schemas/ThresholdConflict" }
        totalFlaggedDonors: { type: integer }
        totalProjectedTouchpoints: { type: integer }
        byThreshold:
          type: array
          items:
            type: object
            required: [thresholdId, thresholdName, flaggedCount]
            properties:
              thresholdId: { type: integer }
              thresholdName: { type: string }
              flaggedCount: { type: integer }
    OverrideInput:
      type: object
      required: [donorIds]
      properties:
        donorIds:
          type: array
          items: { type: string }
    Suppression:
      type: object
      required: [id, campaignId, scope, donorIdCount, createdAt]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        scope: { type: string, enum: [all, channel, campaign_type, touch] }
        channelId: { type: ["integer", "null"] }
        campaignTypeId: { type: ["integer", "null"] }
        touchId: { type: ["integer", "null"] }
        reason: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        donorIdCount: { type: integer }
        createdAt: { type: string, format: date-time }
    SuppressionInput:
      type: object
      required: [scope]
      properties:
        scope: { type: string, enum: [all, channel, campaign_type, touch] }
        channelId: { type: integer }
        campaignTypeId: { type: integer }
        touchId: { type: integer }
        reason: { type: string }
        notes: { type: string }
        rawText: { type: string }
        googleSheetUrl: { type: string }
...
    AppSettings:
      type: object
      required: [fiscalYearStartMonth, fiscalYearStartDay, googleSheetImportEnabled, retentionDeleteEnabled, globalThresholdsEnabled]
      properties:
        fiscalYearStartMonth: { type: integer, minimum: 1, maximum: 12 }
        fiscalYearStartDay: { type: integer, minimum: 1, maximum: 31 }
        googleSheetImportEnabled: { type: boolean }
        retentionDeleteEnabled: { type: boolean }
        globalThresholdsEnabled: { type: boolean }
    SettingsUpdate:
      type: object
      properties:
        fiscalYearStartMonth: { type: integer, minimum: 1, maximum: 12 }
        fiscalYearStartDay: { type: integer, minimum: 1, maximum: 31 }
        googleSheetImportEnabled: { type: boolean }
        retentionDeleteEnabled: { type: boolean }
        globalThresholdsEnabled: { type: boolean }
    RetentionInput:
      type: object
      required: [olderThan]
      properties:
        olderThan: { type: string, format: date }
        confirm: { type: boolean }
    RetentionResult:
      type: object
      required: [campaignsDeleted, touchpointsDeleted]
      properties:
        campaignsDeleted: { type: integer }
        touchpointsDeleted: { type: integer }